requirements.txt
gradio>=4.12
llama-cpp-python>=0.2.24
